fix(ui): refactor and fix user component
