+++ src/app/components/user-me/user-me.component.html
@@ -1,0 +1,138 @@
+<div fxLayout="row wrap">
+  <form fxFlex="100" fxLayout="row wrap" fxLayoustAlign="start stretch" [formGroup]="profileForm">
+    <span fxFlex="100" fxLayout="row wrap" fxLayoutAlign="center center" fxLayoutGap="15px">
+      <mat-form-field>
+        <mat-icon
+          matPrefix
+          *ngIf="!firebaseUser?.emailVerified"
+          title="your email is not verified, click here to request a verification link"
+          (click)="resendVerificationEmail()"
+        >
+          send
+        </mat-icon>
+
+        <mat-icon
+          matPrefix
+          class="material-icons"
+          *ngIf="firebaseUser?.emailVerified"
+          matTooltip="your email is verified"
+        >
+          check
+        </mat-icon>
+
+        <input matInput type="email" formControlName="email" placeholder="your_email@domain.tld" />
+
+        <mat-icon matSuffix class="material-icons"> email </mat-icon>
+
+        <mat-error *ngIf="profileForm.controls.email.invalid">
+          This field is required, and it should be an email
+        </mat-error>
+      </mat-form-field>
+
+      <mat-form-field>
+        <input matInput type="text" formControlName="name" placeholder="your display name" />
+
+        <mat-icon matSuffix class="material-icons"> account_circle </mat-icon>
+      </mat-form-field>
+    </span>
+
+    <span fxFlex="100" fxLayout="row wrap" fxLayoutAlign="center center" fxLayoutGap="15px">
+      <button
+        mat-button
+        *ngIf="!(editMode$ | async)"
+        (click)="resetPassword()"
+        aria-label="reset password"
+      >
+        reset password
+      </button>
+
+      <button mat-button (click)="toggleEditMode()" aria-label="toggle edit mode">
+        <span *ngIf="!(editMode$ | async)">edit profile</span>
+        <span *ngIf="editMode$ | async">cancel</span>
+      </button>
+
+      <button
+        mat-button
+        type="submit"
+        *ngIf="editMode$ | async"
+        [disabled]="profileForm.invalid"
+        (click)="updateProfile()"
+        aria-label="update profile"
+      >
+        update
+      </button>
+
+      <button
+        mat-button
+        type="submit"
+        *ngIf="editMode$ | async"
+        [disabled]="!profileForm.controls.password.value.trim()"
+        (click)="deleteProfile()"
+        color="warn"
+        aria-label="delete profile"
+      >
+        delete
+      </button>
+
+      <mat-form-field *ngIf="editMode$ | async">
+        <input
+          matInput
+          type="{{ (showPassword$ | async) ? 'text' : 'password' }}"
+          formControlName="password"
+          name="password"
+          placeholder="provide a password to DELETE your account"
+        />
+
+        <mat-icon
+          matSuffix
+          class="material-icons"
+          *ngIf="!(showPassword$ | async)"
+          (click)="togglePasswordVisibility()"
+        >
+          visibility
+        </mat-icon>
+
+        <mat-icon
+          matSuffix
+          class="material-icons"
+          *ngIf="showPassword$ | async"
+          (click)="togglePasswordVisibility()"
+        >
+          visibility_off
+        </mat-icon>
+      </mat-form-field>
+
+      <button
+        mat-button
+        *ngIf="!(editMode$ | async) && !me"
+        color="warn"
+        (click)="scConnect()"
+        aria-label="connect with Soundcloud"
+        title="Connect with Soundcloud"
+      >
+        connect with soundcloud
+      </button>
+
+      <button
+        mat-button
+        *ngIf="!(editMode$ | async) && me"
+        color="primary"
+        (click)="scConnect()"
+        aria-label="reconnect with Soundcloud or connect another account"
+        title="Reconnect with Soundcloud or connect another account"
+      >
+        reconnect with soundcloud
+      </button>
+    </span>
+  </form>
+
+  <div fxFlex="100" fxLayout="row wrap">
+    <span fxFlex="100"> Username: {{ me?.username }} </span>
+
+    <span fxFlex="100"> Full name: {{ me?.full_name }} </span>
+
+    <span fxFlex="100"> Playlists: {{ me?.playlist_count }} </span>
+
+    <span fxFlex="100" fxShow fxHide.lt-md [innerHtml]="me?.description"></span>
+  </div>
+</div>
