+++ src/app/components/user-me/user-me.component.scss
@@ -1,0 +1,5 @@
+:host {
+  display: block;
+  width: 100%;
+  height: auto !important;
+}
